simplify sct_dmri_moco

+++ dmri/sct_dmri_moco/j_mri_moco_v8.m
@@ -304,15 +304,20 @@ for indice_index = 1:length(param.index)
         
         % If slice regulation
         if strcmp(todo,'estimate') && strcmp(program,'ANTS')
-            j_disp(fname_log,['Process with ANTS'])
-            mat_tmp=[folder_mat,'mat.T',num2str(iT),'_tmp']; 
-            if ~exist([folder_mat 'nifti_reg'],'dir'), mkdir([folder_mat 'nifti_reg']); end
-            out= [folder_mat 'nifti_reg' filesep num2str(iT) '.nii'];
-            cmd = ['isct_antsSliceRegularizedRegistration -p 2 --output [' mat_tmp ', ' out '] --transform Translation[0.1] --metric MeanSquares[ ' fname_target '.nii* , ' fname_data_splitT_num{iT} '.nii*  , 1 , 16 , Regular , 0.2 ] --iterations 20 -f 1 -s 2'];
-            j_disp(fname_log,['>> ',cmd]);
-            [status, result] = unix(cmd); if status, error(result); end
-            unix(['rm ' out ' ' mat_tmp 'W* ' mat_tmp 'I*'])
-            mat_tmp=[mat_tmp 'TxTy_poly.csv'];
+            if nz>1
+                
+                j_disp(fname_log,['Process with ANTS'])
+                mat_tmp=[folder_mat,'mat.T',num2str(iT),'_tmp'];
+                if ~exist([folder_mat 'nifti_reg'],'dir'), mkdir([folder_mat 'nifti_reg']); end
+                out= [folder_mat 'nifti_reg' filesep num2str(iT) '.nii'];
+                cmd = ['isct_antsSliceRegularizedRegistration -p 2 --output [' mat_tmp ', ' out '] --transform Translation[0.1] --metric MeanSquares[ ' fname_target '.nii* , ' fname_data_splitT_num{iT} '.nii*  , 1 , 16 , Regular , 0.2 ] --iterations 20 -f 1 -s 0.2'];
+                j_disp(fname_log,['>> ',cmd]);
+                [status, result] = unix(cmd); if status, error(result); end
+                unix(['rm ' out ' ' mat_tmp 'W* ' mat_tmp 'I*'])
+                mat_tmp=[mat_tmp 'TxTy_poly.csv'];
+            else 
+                program = 'FLIRT';
+            end
         end
